+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add Stack support
 - Add key generation functionality
 - Add remove key functionality
+- Add clear sign functionality
 - Add progress callback functionality
 
 ## 0.4.0.0